FIx convertDate to use pattern
