+++ src/types/signup.d.ts
@@ -1,0 +1,7 @@
+type SignUpInfoType = {
+  email: string;
+  password: string;
+  passwordCheck: string;
+  name: string;
+  birthDate: string;
+};
