+++ src/utils/regex.ts
@@ -1,0 +1,6 @@
+// HJ: 정규식 패턴들은 파일에 따로 뺴놓는게 관리하기 좋습니당!
+export const emailRegex =
+  /^[0-9a-zA-Z]([-_\\.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_\\.]?[0-9a-zA-Z])*\.[a-zA-Z]{2,3}$/;
+
+export const passwordRegex =
+  /^(?=.*[A-Za-z])(?=.*\d)(?=.*[@$!%*#?&])[A-Za-z\d@$!%*#?&]{8,}$/;
